Conflicts:
	bin/it/polito/oma/etp/solver/ETPsolver_OMAAL_group15.class
	bin/it/polito/oma/etp/solver/Solution.class
	bin/it/polito/oma/etp/solver/TabuInitialization.class
	bin/it/polito/oma/etp/solver/TabuSearch.class
	src/it/polito/oma/etp/solver/ETPsolver_OMAAL_group15.java
	src/it/polito/oma/etp/solver/Solution.java
	src/it/polito/oma/etp/solver/TabuInitialization.java
	src/it/polito/oma/etp/solver/TabuSearch.java
	src/it/polito/oma/etp/solver/TsSettings.java

Thanks to Fabrizio

+++ src/it/polito/oma/etp/solver/initialization/InitializationSolution.java
@@ -33,7 +33,7 @@ public class InitializationSolution extends Solution {
 
 	@Override
 	/*TODO has to be private*/public void initializeFitness() {
-		fitness = penalizingPairs.size();
+		fitness = Float.valueOf(penalizingPairs.size());
 	}
 	
 	/**
